removed deposit steps
